requirements.txt
openai~=1.60.0
python-dotenv~=1.0.1
langchain~=0.3.15